backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
google-generativeai==0.7.2
google-cloud-firestore==2.16.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.12.5
python-multipart==0.0.9
Pillow==12.2.0
pytest==8.2.0
